+++ Blackboard.h
@@ -10,7 +10,7 @@ class Blackboard: public QGraphicsView, public IStreamWR
 {
     Q_OBJECT
 protected:
-    float _scaleRatio;
+    qreal _scaleRatio;
 
     QSize _orginalSize;
 
@@ -79,7 +79,7 @@ public:
 
     QString canvasId() const;
 
-    float scaleRatio();
+    qreal scaleRatio();
 
     QSize orginalSize();
 
@@ -91,7 +91,7 @@ public:
 
     void setOrginalSize(const QSize &size);
 
-    float orginalRatio();
+    qreal orginalRatio();
 
     QSize canvasSize();
